+++ projects/igniteui-angular/directives/src/directives/toggle/toggle.directive.ts
@@ -444,6 +444,10 @@ export class IgxToggleActionDirective implements OnInit {
      * <div igxToggleAction [igxToggleOutlet]="outlet"></div>
      * ```
      * Where `outlet` in an instance of `IgxOverlayOutletDirective` or an `ElementRef`
+     *
+     * @deprecated in version 21.2.0. Overlays now use the HTML Popover API and no longer move to the document
+     * body by default, so using outlet is also no longer needed - just define the overlay in the intended
+     * DOM tree position instead or use `container` property instead.
      */
     @Input('igxToggleOutlet')
     public outlet: IgxOverlayOutletDirective | ElementRef;