+++ projects/igniteui-angular/grids/grid/src/grid-base.directive.ts
@@ -1150,8 +1150,8 @@ export abstract class IgxGridBaseDirective implements GridType,
     /**
      * @hidden @internal
      */
-    @ViewChild('igxLoadingOverlayOutlet', { read: IgxOverlayOutletDirective, static: true })
-    public loadingOutlet: IgxOverlayOutletDirective;
+    @ViewChild('igxLoadingOverlayOutlet', { static: true })
+    public loadingOutlet: ElementRef<HTMLElement>;
 
     /* reactContentChildren */
     /* blazorInclude */
@@ -4491,6 +4491,10 @@ export abstract class IgxGridBaseDirective implements GridType,
      *
      * @remarks
      * If set, returns the outlet defined outside the grid. Otherwise returns the grid's internal outlet directive.
+     *
+     * @deprecated in version 21.2.0. Overlays now use the HTML Popover API and no longer move to the document
+     * body by default, so using outlet is also no longer needed - just define the overlay in the intended
+     * DOM tree position instead or use `container` property instead.
      */
     @Input()
     public get outlet() {
@@ -6205,6 +6209,7 @@ export abstract class IgxGridBaseDirective implements GridType,
      */
     public showSnackbarFor(index: number) {
         this.addRowSnackbar.actionText = index === -1 ? '' : this.resourceStrings.igx_grid_snackbar_addrow_actiontext;
+        this.addRowSnackbar.positioning = 'container';
         this.lastAddedRowIndex = index;
         this.addRowSnackbar.open();
     }